lang_translation/requirements.txt
django==4.2.4
googletrans==4.0.0-rc1
djangorestframework==3.14.0